+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Support multi individual input VCFs for calling
+### Changed
+- Output FORMAT tags to allow sample resolution on multi individual VCFs
+
 ## [0.2]
 ### Added
 - Build reference, use prebuilt reference